refactor and use a project route
