+++ middleware/auth.js
@@ -1,0 +1,9 @@
+// User Authentication middleware
+function ensureAuthenticated(req, res, next) {
+    if (req.isAuthenticated()) {
+        return next();
+    }
+    res.redirect("/login");
+}
+
+module.exports = ensureAuthenticated;
